Github support: TEXT mode and UTF-8 support for all objects (#58)
As discussed in #25, add support to work with a plain folder, rather than SVN. This allows the use of any SCC system (including git). This is refered to as TEXT mode (in contrast to SVN mode).

In addition there is now support for encoding all files as UTF-8, rather than just UCS2 encoded files. This helps when working with SCC tools that expect UTF-8 (rather than CP1252).

These changes make it possible to the the add-in to develop projects hosted on GitHub.

Co-authored-by: Barry Pitfield <barry@x-ware.co.uk>
Co-authored-by: Oliver <oliver.vahey@talkjcs.com>