Fallback to finding references in Distributables
If ref cannon be found by GUID or FullPath, try looking in
<Import Path>\Distributables\<Ref File Name>

The project that is built won't actually be able to run, as the ref will not
be able to resolve at runtime if it hasn't been registered with RegSvr32.

This allows you to have a generic Access build computer, that doesn't need any
particular prerequisite libraries installed.