Automatically re-import references if any are broken (#68)
Fix references on import. This allows compiling the database on a machine that
does not have References installed/registered ( when libraries are installed in
`<import path>/Distributables`). This is a fix/completion of the work done in
#64.

Also protect against the FullPath of a Reference not being available. This
happens if a Reference is not registered.